moved classes for proper implementation
