+++ Risea/Src/MSComponent.cpp
@@ -1,0 +1,8 @@
+#include "MSComponent.h"
+
+MSComponent::MSComponent() {
+
+}
+MSComponent::~MSComponent(){
+
+}
