+++ Risea/Src/MSComponent.h
@@ -19,11 +19,12 @@ class MSComponent : public juce::Component
 {
     public:
     MSComponent();
+    ~MSComponent();
 
-    void paint(juce::Graphics&) override;
+    /*void paint(juce::Graphics&) override;
     void resized() override;
     void drawButtonBackground(juce::Graphics&, juce::Button&, const juce::Colour&, bool);
-
+*/
     private:
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (MSComponent)
